shore(FE): restructure hooks folder
