+++ frontend/src/api/hooks/restaurant/queries/useRestaurants.ts
@@ -1,5 +1,5 @@
 import { RestaurantResponse } from '../../../../types/menu-item'
-import useBaseQuery from '../useBaseQuery'
+import useBaseQuery from '../../useBaseQuery'
 
 const useRestaurants = () => useBaseQuery<RestaurantResponse>({
   apiParams: { path: '/restaurant' },
