+++ frontend/src/pages/Register.page.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react'
 import { toast } from 'react-toastify'
 
-import useRegisterMutation from '../api/hooks/mutations/auth/useRegisterMutation'
+import useRegisterMutation from '../api/hooks/auth/mutations/useRegisterMutation'
 import useValidation from '../hooks/useValidation'
 import { trimObjectStrings } from '../utils/string.utils'
 import { ERROR } from '../utils/error-message.constants'
